requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
certifi==2020.12.5
chardet==3.0.4
decorator==4.4.2
distlib==0.3.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fett==0.3.2
filelock==3.0.12
idna==2.10
networkx==2.5
pathtools==0.1.2
pymongo==3.11.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
snooty-lextudio==1.8.6.dev0
soupsieve==2.0.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.2.0
watchdog==0.10.4
